Both OpenGL funcitons and Shader are failing to establish depth for sprites ATM
OpenGL is probably failing for the simple reason that I need to use OpenGL's texture, transforms and matrix to make use of drawing depth with their functions. The shader not working is odd , given the fact that I'm just sending the sprites Z value into the shader and then using the shader in the drawing call of every sprite.