Depth is now established by drawing everything inside a list and QuickSorting it
It's ugly and bad for performance, but the only way to have a dynamic orthographic perspective in SFML without this method is by using OpenGL from scratch to draw textures. Only Sprites that are in the view of the Camera are added to the list and drawn. The only upside to this is that you can call a Sprite to be drawn from anywhere, not just the Draw() method.

Next goal in Sprite-handling is probably making the Sprite or Texture inside the GSprite class to be a pointer that points to already existing Sprites that are have been previously loaded in the game(they would be stored inside a map/Dictionary that has the file path as the key).

Other noteable goals right now are establishing a system for Collision Checking. I'm going to research the use of QuadTree to check Collision.