Started to implement AStar Pathfinding. Memory issue is worse than previously thought
The memory leak is not based on sf::Texture but instead almost every single object that gets created, wether on the Stack or the Heap, which is pretty alarming to say the least. Though it seems to be objects outside of std namespace(anything created for the C++ library).

I thought I could fix it but it seems like that's not true, and my understanding of the problem has only got more mumbled and less understandable since the start. Probably need some help on this one.