Sprites for enemies, Projectile Enemy close to done.
Thinking of doing a different way to draw out sprites, instead of putting them in a list, sorting them, drawing them and then removing them from the list there would be an entity list that would always be sorted by their depth value. This would lead to being able to draw a sprite more than once per frame. To optimize there would be a map of sf::sprites based on their texture destination(on drive) so that no sprite or textures would need to be loaded twice but can instead be pointed to. Just a though though.